--- old_webpack_stuff/.gitignore
@@ -1,11 +1,0 @@
-#############
-## Node
-#############
-
-node_modules/
-
-#############
-## Generated Build Files
-#############
-
-build/
