--- old_webpack_stuff/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "stk-toolkit",
-  "version": "4.0.0",
-  "description": "{ Project }",
-  "main": "index.js",
-  "scripts": {
-    "start": "node_modules/.bin/nodemon --exec babel-node --debug webpack/production.js --watch webpack/production.js",
-    "dev": "node_modules/.bin/babel-node webpack/development.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/stoikerty/stk-toolkit.git"
-  },
-  "keywords": [
-    "webpack",
-    "sass",
-    "react",
-    "browserify",
-    "gulp",
-    "moonboots",
-    "middleman"
-  ],
-  "author": "Jorge Antunes",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/stoikerty/stk-toolkit/issues"
-  },
-  "homepage": "https://github.com/stoikerty/stk-toolkit#readme",
-  "devDependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-watch": "^1.0.1",
-    "nodemon": "^1.9.1",
-    "webpack": "^1.12.14"
-  },
-  "dependencies": {
-    "domready": "^1.0.8",
-    "express": "^4.13.4",
-    "express-handlebars": "^3.0.0",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-router": "^2.0.0-rc4",
-    "serve-static": "^1.10.2"
-  }
-}
