--- old_webpack_stuff/src/server/views/layout.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <!-- Document Settings -->
-  <meta http-equiv="Content-Type" content="text/html" charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-
-   <!-- Page Meta -->
-  <title>@stoikerty</title>
-  <meta name="description" content="@stoikerty - Jorge Antunes" />
-
-  <!-- Responsive Meta Tags -->
-  <meta name="HandheldFriendly" content="True" />
-  <meta name="MobileOptimized" content="320" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-  <!-- Use Financial Times polyfill service http://cdn.polyfill.io/ -->
-  <script src="https://cdn.polyfill.io/v1/polyfill.min.js"></script>
-
-  <!-- Fonts -->
-  <link href='https://fonts.googleapis.com/css?family=Source+Sans+Pro:200,300,400,600,700,900' rel='stylesheet' type='text/css'>
-  <!-- Icons -->
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
-
-  <link rel="apple-touch-icon" sizes="57x57" href="favicon-files/apple-touch-icon-57x57.png?v=2">
-  <link rel="apple-touch-icon" sizes="60x60" href="favicon-files/apple-touch-icon-60x60.png?v=2">
-  <link rel="apple-touch-icon" sizes="72x72" href="favicon-files/apple-touch-icon-72x72.png?v=2">
-  <link rel="apple-touch-icon" sizes="76x76" href="favicon-files/apple-touch-icon-76x76.png?v=2">
-  <link rel="apple-touch-icon" sizes="114x114" href="favicon-files/apple-touch-icon-114x114.png?v=2">
-  <link rel="apple-touch-icon" sizes="120x120" href="favicon-files/apple-touch-icon-120x120.png?v=2">
-  <link rel="apple-touch-icon" sizes="144x144" href="favicon-files/apple-touch-icon-144x144.png?v=2">
-  <link rel="apple-touch-icon" sizes="152x152" href="favicon-files/apple-touch-icon-152x152.png?v=2">
-  <link rel="apple-touch-icon" sizes="180x180" href="favicon-files/apple-touch-icon-180x180.png?v=2">
-  <link rel="icon" type="image/png" href="favicon-files/favicon-32x32.png?v=2" sizes="32x32">
-  <link rel="icon" type="image/png" href="favicon-files/android-chrome-192x192.png?v=2" sizes="192x192">
-  <link rel="icon" type="image/png" href="favicon-files/favicon-96x96.png?v=2" sizes="96x96">
-  <link rel="icon" type="image/png" href="favicon-files/favicon-16x16.png?v=2" sizes="16x16">
-  <link rel="manifest" href="favicon-files/manifest.json?v=3">
-  <link rel="shortcut icon" href="favicon-files/favicon.ico?v=2">
-  <meta name="msapplication-TileColor" content="#da532c">
-  <meta name="msapplication-TileImage" content="favicon-files/mstile-144x144.png?v=2">
-  <meta name="msapplication-config" content="favicon-files/browserconfig.xml?v=2">
-  <meta name="theme-color" content="#ffffff">
-
-  <script src="app.js"></script>
-</head>
-<html>
-  <body>
-    <div data-jshook="app-body">{{{reactHtml}}}</div>
-  </body>
-</html>
