+++ package.json
@@ -28,13 +28,15 @@
   "homepage": "https://github.com/stoikerty/stk-toolkit#readme",
   "devDependencies": {
     "babel-cli": "^6.6.5",
+    "babel-core": "^6.7.2",
     "babel-loader": "^6.2.4",
-    "babel-plugin-transform-runtime": "^6.6.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
-    "babel-watch": "^1.0.1",
     "nodemon": "^1.9.1",
-    "webpack": "^1.12.14"
+    "redbox-react": "^1.2.2",
+    "webpack": "^1.12.14",
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-hot-middleware": "^2.10.0"
   },
   "dependencies": {
     "domready": "^1.0.8",
