--- redux_example/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Redux async example</title>
-  </head>
-  <body>
-    <div id="root">
-    </div>
-    <script src="/static/bundle.js"></script>
-  </body>
-</html>
