--- redux_example/index.js
@@ -1,43 +1,0 @@
-import 'babel-polyfill'
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Provider } from 'react-redux'
-import configureStore from './store/configureStore'
-
-const store = configureStore()
-const rootEl = document.getElementById('root')
-
-let render = () => {
-  const App = require('./containers/App').default
-  ReactDOM.render(
-    <Provider store={store}>
-      <App />
-    </Provider>,
-    rootEl
-  )
-}
-
-if (module.hot) {
-  // Support hot reloading of components
-  // and display an overlay for runtime errors
-  const renderApp = render
-  const renderError = (error) => {
-    const RedBox = require('redbox-react')
-    ReactDOM.render(
-      <RedBox error={error} />,
-      rootEl
-    )
-  }
-  render = () => {
-    try {
-      renderApp()
-    } catch (error) {
-      renderError(error)
-    }
-  }
-  module.hot.accept('./containers/App', () => {
-    setTimeout(render)
-  })
-}
-
-render()
