--- redux_example/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "redux-async-example",
-  "version": "0.0.0",
-  "description": "Redux async example",
-  "scripts": {
-    "start": "node server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/reactjs/redux.git"
-  },
-  "keywords": [
-    "react",
-    "reactjs",
-    "hot",
-    "reload",
-    "hmr",
-    "live",
-    "edit",
-    "webpack",
-    "flux"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/reactjs/redux/issues"
-  },
-  "homepage": "http://redux.js.org",
-  "dependencies": {
-    "babel-polyfill": "^6.3.14",
-    "isomorphic-fetch": "^2.1.1",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-redux": "^4.2.1",
-    "redux": "^3.2.1",
-    "redux-logger": "^2.4.0",
-    "redux-thunk": "^1.0.3"
-  },
-  "devDependencies": {
-    "babel-core": "^6.3.15",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "expect": "^1.6.0",
-    "express": "^4.13.3",
-    "node-libs-browser": "^0.5.2",
-    "redbox-react": "^1.2.2",
-    "webpack": "^1.9.11",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.2.0"
-  }
-}
